Refactor Picture Loader.
