+++ cockatrice/src/dialogs/dlg_manage_sets.cpp
@@ -1,7 +1,7 @@
 #include "dlg_manage_sets.h"
 
 #include "../client/network/sets_model.h"
-#include "../client/ui/picture_loader.h"
+#include "../client/ui/picture_loader/picture_loader.h"
 #include "../deck/custom_line_edit.h"
 #include "../game/cards/card_database_manager.h"
 #include "../main.h"
