+++ cockatrice/src/dialogs/dlg_settings.cpp
@@ -4,7 +4,7 @@
 #include "../client/network/release_channel.h"
 #include "../client/network/spoiler_background_updater.h"
 #include "../client/sound_engine.h"
-#include "../client/ui/picture_loader.h"
+#include "../client/ui/picture_loader/picture_loader.h"
 #include "../client/ui/theme_manager.h"
 #include "../deck/custom_line_edit.h"
 #include "../game/cards/card_database.h"
